+++ src/components/library/AlbumView.tsx
@@ -219,15 +219,14 @@ const AlbumView = ({ ...rest }: any) => {
             subtitle={
               <div>
                 <PageHeaderSubtitleDataLine $top $overflow>
-                  <StyledLink color="#D8D8D8" onClick={() => history.push('/library/album')}>
-                    {t('ALBUM')}
+                  <StyledLink onClick={() => history.push('/library/album')}>
+                    {t('Album')}
                   </StyledLink>{' '}
                   {data.albumArtist && (
                     <>
                       {t('by')}{' '}
                       <LinkWrapper maxWidth="20vw">
                         <StyledLink
-                          color="#D8D8D8"
                           onClick={() => history.push(`/library/artist/${data.albumArtistId}`)}
                         >
                           <strong>{data.albumArtist}</strong>
@@ -265,7 +264,6 @@ const AlbumView = ({ ...rest }: any) => {
                         {i > 0 && ', '}
                         <LinkWrapper maxWidth="13vw">
                           <StyledLink
-                            color="#D8D8D8"
                             tabIndex={0}
                             onClick={() => {
                               if (!rest.isModal) {